+++ src/Utilities/Camera.hpp
@@ -6,7 +6,7 @@ class Camera {
 public:
     Camera(Vec3 position, Vec3 target, Vec3 up,
            float viewportWidth, float viewportHeight);
-    Ray rayForPixel(int x, int y, int imageWidth, int imageHeight) const;
+    Ray rayForPixel(float x, float y, int imageWidth, int imageHeight) const;
     const Vec3& position() const { return position_; }
 
 private:
